requirements.txt
openai==0.26.5
pandas==1.5.3
python-dotenv==1.0.0
